Remove Maven plugin for building Docker images
Gets rid of the Spotify Docker plugin we were using to produce Elucidate Docker
images and replaces it with a multi-stage build Dockerfile that can be used to
build a new container from sources.

Additionally, fixes a bug where the token signing key was still being
configured even when authentication was disabled.